pyproject.toml
[project]
name = "ai-stock-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.19",
]
